requirements.txt
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
yfinance>=0.2.18
streamlit>=1.28.0
plotly>=5.15.0
requests>=2.31.0
scipy>=1.11.0
kafka-python>=2.0.2
redis>=5.0.0
matplotlib>=3.7.0
networkx>=3.1
